Fix problem matcher config
When using the built-in build command (ctrl+shift+b, or when starting a
debug session) vscode will parse the output for "problems". This was
working but trying to navigate to the issue fails.

This fixes the config to ensure paths are interpreted correctly, and to
also make the build task compile everything instead of stopping on
error.