Add support for dot matrix board to wscript
Add support for dot matrix board to wscript. Using the --board option
while running waf configure allows the user to set the default board to
DOT_MATRIX or ASCII.

Currently the captouch support for the dot matrix board is disabled due
to the Atmel QTouch library for GCC being missing.