Ignore toolchains that have large versions
Seen on one student's setup, a version of GCC (Maybe linaroo?) that used
the year of the release as the major version number. Since this is very
large it was being selected in preference to the toolchain installed
for EiE, and could not parse the linker.

To fix this for now, lets just ignore anything with a major version
that looks like a year.